Revert "Add exception for default rooms beta for users who are members of a free policy"
