+++ src/libs/Navigation/AppNavigator/MainDrawerNavigator.js
@@ -13,8 +13,6 @@ import ReportScreen from '../../../pages/home/ReportScreen';
 import SidebarScreen from '../../../pages/home/sidebar/SidebarScreen';
 import BaseDrawerNavigator from './BaseDrawerNavigator';
 import * as ReportUtils from '../../ReportUtils';
-import * as PolicyUtils from '../../PolicyUtils';
-import CONST from '../../../CONST';
 
 const propTypes = {
     /** Available reports that would be displayed in this navigator */
@@ -24,15 +22,6 @@ const propTypes = {
 
     /** Beta features list */
     betas: PropTypes.arrayOf(PropTypes.string),
-
-    /** The policies which the user has access to */
-    policies: PropTypes.objectOf(PropTypes.shape({
-        /** The policy name */
-        name: PropTypes.string,
-
-        /** The type of the policy */
-        type: PropTypes.string,
-    })).isRequired,
 };
 
 const defaultProps = {
@@ -44,11 +33,11 @@ const defaultProps = {
  * Get the most recently accessed report for the user
  *
  * @param {Object} reports
- * @param {String[]} [reportTypesToIgnore]
+ * @param {Boolean} [ignoreDefaultRooms]
  * @returns {Object}
  */
-const getInitialReportScreenParams = (reports, reportTypesToIgnore) => {
-    const last = ReportUtils.findLastAccessedReport(reports, reportTypesToIgnore);
+const getInitialReportScreenParams = (reports, ignoreDefaultRooms) => {
+    const last = ReportUtils.findLastAccessedReport(reports, ignoreDefaultRooms);
 
     // Fallback to empty if for some reason reportID cannot be derived - prevents the app from crashing
     const reportID = lodashGet(last, 'reportID', '');
@@ -56,20 +45,7 @@ const getInitialReportScreenParams = (reports, reportTypesToIgnore) => {
 };
 
 const MainDrawerNavigator = (props) => {
-    // If one is a member of a free policy, then they are allowed to see the Policy default rooms.
-    // For everyone else, one must be on the beta to see a default room.
-    let reportTypesToIgnore = [];
-    const isMemberOfFreePolicy = PolicyUtils.isMemberOfFreePolicy(props.policies);
-    if (isMemberOfFreePolicy && !Permissions.canUseDefaultRooms(props.betas)) {
-        reportTypesToIgnore = [CONST.REPORT.CHAT_TYPE.DOMAIN_ALL];
-    } else if (!Permissions.canUseDefaultRooms(props.betas)) {
-        reportTypesToIgnore = [
-            CONST.REPORT.CHAT_TYPE.POLICY_ADMINS,
-            CONST.REPORT.CHAT_TYPE.POLICY_ANNOUNCE,
-            CONST.REPORT.CHAT_TYPE.DOMAIN_ALL,
-        ];
-    }
-    const initialParams = getInitialReportScreenParams(props.reports, reportTypesToIgnore);
+    const initialParams = getInitialReportScreenParams(props.reports, !Permissions.canUseDefaultRooms(props.betas));
 
     // Wait until reports are fetched and there is a reportID in initialParams
     if (!initialParams.reportID) {
@@ -104,9 +80,6 @@ export default withOnyx({
     },
     betas: {
         key: ONYXKEYS.BETAS,
-    },
-    policies: {
-        key: ONYXKEYS.COLLECTION.POLICY,
     },
 })(MainDrawerNavigator);
 export {getInitialReportScreenParams};
