--- src/libs/PolicyUtils.js
@@ -1,30 +1,0 @@
-import _ from 'underscore';
-import CONST from '../CONST';
-
-/**
- * Is the user an admin of a free policy (aka workspace)?
- *
- * @param {Array} policies
- * @returns {Boolean}
- */
-function isAdminOfFreePolicy(policies) {
-    return _.some(policies, policy => policy
-        && policy.type === CONST.POLICY.TYPE.FREE
-        && policy.role === CONST.POLICY.ROLE.ADMIN);
-}
-
-/**
- * Is the user a member of a free policy (aka workspace)?
- *
- * @param {Array} policies
- * @returns {Boolean}
- */
-function isMemberOfFreePolicy(policies) {
-    return _.some(policies, policy => policy
-        && policy.type === CONST.POLICY.TYPE.FREE);
-}
-
-export {
-    isAdminOfFreePolicy,
-    isMemberOfFreePolicy,
-};
