+++ src/libs/actions/Policy.js
@@ -268,6 +268,18 @@ function loadFullPolicy(policyID) {
 }
 
 /**
+ * Is the user an admin of a free policy (aka workspace)?
+ *
+ * @param {Array} policies
+ * @returns {Boolean}
+ */
+function isAdminOfFreePolicy(policies) {
+    return _.some(policies, policy => policy
+        && policy.type === CONST.POLICY.TYPE.FREE
+        && policy.role === CONST.POLICY.ROLE.ADMIN);
+}
+
+/**
  * Remove the passed members from the policy employeeList
  *
  * @param {Array} members
@@ -538,6 +550,7 @@ export {
     loadFullPolicy,
     removeMembers,
     invite,
+    isAdminOfFreePolicy,
     create,
     uploadAvatar,
     update,
