+++ src/libs/actions/Welcome.js
@@ -4,11 +4,11 @@ import lodashGet from 'lodash/get';
 import Navigation from '../Navigation/Navigation';
 import * as ReportUtils from '../ReportUtils';
 import ROUTES from '../../ROUTES';
+import * as Policy from './Policy';
 import ONYXKEYS from '../../ONYXKEYS';
 import NameValuePair from './NameValuePair';
 import CONST from '../../CONST';
 import SCREENS from '../../SCREENS';
-import * as PolicyUtils from '../PolicyUtils';
 
 let resolveIsReadyPromise;
 let isReadyPromise = new Promise((resolve) => {
@@ -130,7 +130,7 @@ function show({routes, showCreateMenu}) {
 
         // If user is not already an admin of a free policy and we are not navigating them to their workspace or creating a new workspace via workspace/new then
         // we will show the create menu.
-        if (!PolicyUtils.isAdminOfFreePolicy(allPolicies) && !isDisplayingWorkspaceRoute) {
+        if (!Policy.isAdminOfFreePolicy(allPolicies) && !isDisplayingWorkspaceRoute) {
             showCreateMenu();
         }
     });
